+++ identity_storage/src/storage/tests/mod.rs
@@ -1,0 +1,3 @@
+// Copyright 2020-2023 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+mod api;
